Fix compilation issues with MSVC 2010 (mostly by eliminating use of range-based for loops and std::tie)

+++ glslang/Include/Common.h
@@ -206,7 +206,7 @@ struct TSourceLoc {
     {
         if (name != nullptr)
             return quoteStringName ? ("\"" + std::string(name) + "\"") : name;
-        return std::to_string(string);
+        return std::to_string((long long)string);
     }
     const char* name; // descriptive name for this string
     int string;
